Refactor GetCryptocurrencyInfo function to use new API endpoint
The GetCryptocurrencyInfo function in cryptocurrency-info.go has been refactored to use the new "/v2/cryptocurrency/info" API endpoint instead of the previous "/v1/cryptocurrency/info" endpoint. This change was made to improve the efficiency and accuracy of retrieving cryptocurrency information. The function now takes an additional parameter, "id", which specifies the ID of the cryptocurrency for which information is requested. The function makes a request to the API with the provided ID and retrieves the corresponding cryptocurrency information. If the API rate limit is exceeded, the function will automatically sleep for 62 seconds before retrying the request. The response is then unmarshaled into a CryptocurrencyInfoResponse struct and returned. Overall, this refactor improves the functionality and performance of the GetCryptocurrencyInfo function.